RestSentiBackend/requirements.txt
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
importlib-metadata==5.0.0
joblib==1.2.0
Markdown==3.4.1
nltk==3.7
numpy==1.23.5
pandas==1.5.1
pickle5==0.0.11
python-dateutil==2.8.2
pytz==2022.6
regex==2022.10.31
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.3
threadpoolctl==3.1.0
tqdm==4.64.1
urllib3==1.26.12
zipp==3.10.0
